Refactor and reorganisation
